// frontend/src/types/api.ts
export interface CartItem {
  id: string;
  type: 'product' | 'service';
  name: string;
  price: number;
  sale_price?: number;
  promo_price?: number;
  image_url?: string;
  quantity: number;
  stock?: number;
  category?: string;
}